adds missing optionnal options to all api calls
As documented in node-mailjet's [README](https://github.com/mailjet/mailjet-apiv3-nodejs), the mailjet client acutally accepts [ConfigOptions](https://github.com/mailjet/mailjet-apiv3-nodejs#client--call-configuration-specifics) for all calls of all APIs. 
This allows to override the client's api version for a specific call among other things. 

I propose this change because we primarily use the `v3.1` send api, but have 2 calls to `v3 `-only apis. This change would allow us to configure `v3.1` at the client level and override the version only for the 2 calls which require `v3` and drop the dozens of version overrides we have in the code base.